+++ crates/rust-client/src/sync/mod.rs
@@ -217,6 +217,8 @@ pub struct SyncSummary {
     pub locked_accounts: Vec<AccountId>,
     /// IDs of committed transactions.
     pub committed_transactions: Vec<TransactionId>,
+    /// IDs of discarded transactions.
+    pub discarded_transactions: Vec<TransactionId>,
 }
 
 impl SyncSummary {
@@ -228,6 +230,7 @@ impl SyncSummary {
         updated_accounts: Vec<AccountId>,
         locked_accounts: Vec<AccountId>,
         committed_transactions: Vec<TransactionId>,
+        discarded_transactions: Vec<TransactionId>,
     ) -> Self {
         Self {
             block_num,
@@ -237,6 +240,7 @@ impl SyncSummary {
             updated_accounts,
             locked_accounts,
             committed_transactions,
+            discarded_transactions,
         }
     }
 
@@ -249,6 +253,7 @@ impl SyncSummary {
             updated_accounts: vec![],
             locked_accounts: vec![],
             committed_transactions: vec![],
+            discarded_transactions: vec![],
         }
     }
 
@@ -259,6 +264,7 @@ impl SyncSummary {
             && self.updated_accounts.is_empty()
             && self.locked_accounts.is_empty()
             && self.committed_transactions.is_empty()
+            && self.discarded_transactions.is_empty()
     }
 
     pub fn combine_with(&mut self, mut other: Self) {
@@ -269,6 +275,7 @@ impl SyncSummary {
         self.updated_accounts.append(&mut other.updated_accounts);
         self.locked_accounts.append(&mut other.locked_accounts);
         self.committed_transactions.append(&mut other.committed_transactions);
+        self.discarded_transactions.append(&mut other.discarded_transactions);
     }
 }
 
@@ -281,6 +288,7 @@ impl Serializable for SyncSummary {
         self.updated_accounts.write_into(target);
         self.locked_accounts.write_into(target);
         self.committed_transactions.write_into(target);
+        self.discarded_transactions.write_into(target);
     }
 }
 
@@ -295,6 +303,7 @@ impl Deserializable for SyncSummary {
         let updated_accounts = Vec::<AccountId>::read_from(source)?;
         let locked_accounts = Vec::<AccountId>::read_from(source)?;
         let committed_transactions = Vec::<TransactionId>::read_from(source)?;
+        let discarded_transactions = Vec::<TransactionId>::read_from(source)?;
 
         Ok(Self {
             block_num,
@@ -304,6 +313,7 @@ impl Deserializable for SyncSummary {
             updated_accounts,
             locked_accounts,
             committed_transactions,
+            discarded_transactions,
         })
     }
 }