+++ tests/src/network_transaction_tests.rs
@@ -5,13 +5,11 @@ use miden_client::{
     account::{Account, AccountBuilder, AccountStorageMode, StorageSlot},
     note::NoteTag,
     testing::{
-        common::{
-            TestClient, create_test_client, execute_tx_and_sync, insert_new_wallet,
-            wait_for_blocks, wait_for_tx,
-        },
+        common::{TestClient, create_test_client, wait_for_tx},
         note::NoteBuilder,
     },
     transaction::{OutputNote, TransactionRequestBuilder, TransactionScript},
+    utils::{execute_tx_and_sync, insert_new_wallet, wait_for_blocks},
 };
 use miden_lib::transaction::TransactionKernel;
 use miden_objects::{
@@ -188,9 +186,9 @@ async fn counter_contract_ntx() {
         .build()
         .unwrap();
 
-    execute_tx_and_sync(&mut client, native_account.id(), tx_request).await;
+    execute_tx_and_sync(&mut client, native_account.id(), tx_request).await.unwrap();
 
-    wait_for_blocks(&mut client, 2).await;
+    wait_for_blocks(&mut client, 2).await.unwrap();
 
     let a = client
         .test_rpc_api()
@@ -267,7 +265,7 @@ async fn recall_note_before_ntx_consumes_it() {
 
     client.testing_apply_transaction(consume_transaction).await.unwrap();
 
-    wait_for_blocks(&mut client, 2).await;
+    wait_for_blocks(&mut client, 2).await.unwrap();
 
     // The network account should have original value
     assert_eq!(